add some comments for OapCheckIndex Class and move its test to OapCheckIndexSuite
